Allow PSTN phone account for incoming calls from connection manager
Currently the phone account specified for an incoming call must
match belong to the calling app.

This CL relaxes that restricition a little. Now connection managers
can specify PSTN phone accounts when adding a new incoming call.

This is necessary so that incoming calls from the connection manager
are displayed the same way as outgoing calls from the connection
manager.

Bug: 19591608, 19217193
Change-Id: I856ada853865c0ca01a2c3f9baede354f0855ea8

Conflicts:
	src/com/android/server/telecom/TelecomService.java